modify daily panel to include the circles which reflect the daily macros and make hte panel scrolable
